move slot advanced prosps from common  to slot edits
