+++ src/components/dashboard/containers/ItemCommon/itemCommonNew.js
@@ -373,176 +373,6 @@ export const MediaCard = ({ mediaUrl = '', mediaMime = '', label = '' }) => {
 	)
 }
 
-export const ItemMinPerImpression = ({
-	item = {},
-	updateMultipleFields,
-	validations,
-	validateId,
-	updateField,
-	setActiveFields,
-	returnPropToInitialState,
-	activeFields = {},
-}) => {
-	const { address, decimals, symbol } = selectMainToken()
-	const { minPerImpression, rulesInput = { inputs: {} } } = item
-	const { autoSetMinCPM } = rulesInput.inputs
-	const active = !!activeFields.minPerImpression
-	const { minPerImpression: errMin } = validations
-	const showError = !!errMin && errMin.dirty
-	const minCPM =
-		typeof minPerImpression === 'object'
-			? formatTokenAmount(
-					bigNumberify((item.minPerImpression || {})[address] || '0').mul(1000),
-					decimals,
-					true
-			  )
-			: minPerImpression
-
-	return (
-		<Grid container spacing={2}>
-			<Grid item xs={12}>
-				<TextField
-					fullWidth
-					variant='outlined'
-					type='text'
-					required
-					label={t('MIN_CPM_SLOT_LABEL_MANUAL')}
-					name='minPerImpression'
-					value={minCPM || ''}
-					disabled={!!autoSetMinCPM || !active}
-					onChange={ev => {
-						const value = ev.target.value
-						updateField('minPerImpression', value, {
-							name: 'advanced',
-							fields: ['rulesInput', 'minPerImpression'],
-						})
-						execute(
-							validateNumberString({
-								validateId,
-								prop: 'minPerImpression',
-								value,
-								dirty: true,
-							})
-						)
-					}}
-					error={showError}
-					maxLength={120}
-					helperText={
-						showError ? errMin.errMsg : t('SLOT_MANUAL_CPM_MIN_HELPER')
-					}
-					InputProps={
-						!autoSetMinCPM
-							? {
-									endAdornment: (
-										<InputAdornment position='end'>
-											<IconButton
-												// size='small'
-												color='secondary'
-												onClick={() =>
-													active
-														? returnPropToInitialState('minPerImpression')
-														: setActiveFields('minPerImpression', true)
-												}
-											>
-												{active ? <UndoOutlined /> : <Edit />}
-											</IconButton>
-										</InputAdornment>
-									),
-							  }
-							: {}
-					}
-				/>
-			</Grid>
-
-			<Grid item xs={12}>
-				<OutlinedPropView
-					label={t('SLOT_AUTO_MIN_CPM_LABEL')}
-					value={
-						<FormControl>
-							<FormGroup row>
-								<FormControlLabel
-									control={
-										<Checkbox
-											checked={!!autoSetMinCPM}
-											onChange={ev => {
-												const checked = ev.target.checked
-												updateField(
-													'rulesInput',
-													{
-														...rulesInput,
-														inputs: {
-															...rulesInput.inputs,
-															autoSetMinCPM: checked,
-														},
-													},
-													{
-														name: 'advanced',
-														fields: ['rulesInput', 'minPerImpression'],
-													}
-												)
-											}}
-											value='autoSetMinCPM'
-										/>
-									}
-									label={t('SLOT_AUTO_MIN_CPM_INFO_LABEL')}
-								/>
-							</FormGroup>
-						</FormControl>
-					}
-				/>
-			</Grid>
-		</Grid>
-	)
-}
-export const SlotAdultContent = ({
-	item = {},
-	// validations,
-	updateField,
-	// setActiveFields,
-	// returnPropToInitialState,
-	// activeFields = {},
-}) => {
-	const { rulesInput = { inputs: {} } } = item
-	const { allowAdultContent } = rulesInput.inputs
-
-	return (
-		<OutlinedPropView
-			label={t('SLOT_ALLOW_ADULT_CONTENT_LABEL')}
-			value={
-				<FormControl>
-					<FormGroup row>
-						<FormControlLabel
-							control={
-								<Checkbox
-									checked={!!allowAdultContent}
-									onChange={ev =>
-										updateField(
-											'rulesInput',
-											{
-												...rulesInput,
-												inputs: {
-													...rulesInput.inputs,
-													allowAdultContent: ev.target.checked,
-												},
-											},
-											{
-												name: 'advanced',
-												fields: ['rulesInput', 'minPerImpression'],
-											}
-										)
-									}
-									value='allowAdultContent'
-								/>
-							}
-							label={t('SLOT_ALLOW_ADULT_CONTENT_INFO_LABEL')}
-						/>
-					</FormGroup>
-				</FormControl>
-			}
-		/>
-	)
-}
-
 export const ItemWebsite = ({
 	item = {},
 	initialItemState = {},
